+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- RAW image data can be loaded as texture.
+- Z-level of 3D-textures can be selected in preview.
 
 ## [Version 1.12] - 2020-09-10
 ### Added